refactor: api 파일위치 및 type 네임 개선
